Adopt the Deck skin, drop its bundle, add the planets
Deck is weewx-wdc v3.5.1, taken over and rewritten onto our own layers.
Rendering it never loads a module called `weewx`. `tags.py` imports from
`weewx_evo.skinkit` by name. The shim that answers `import weewx` is
installed only for a skin from outside.

What the pages loaded before, per view:

    1.7 MB  React and nivo, in a bundle whose TypeScript is not here
    1.0 MB  Plotly, for one chart
    658 KB  Carbon
    556 KB  fonts
        21  script tags to IBM's CDN, in five versions

Two of those versions defined `bx-tooltip`. Every page logged
"Attempting to re-define bx-tooltip".

Now: ECharts, one stylesheet, two scripts. 3.9 MB to 1.16 MB, 21 foreign
requests to none, three chart libraries to one. `tools/deck_test.py`
renders the skin and reads the pages back. It fails
on a Carbon class, a foreign fetch, an American date on a German page, a
US unit on a metric one, an empty chart, or a tab pointing at nothing.

The data bindings are gone. WeeWX reaches a second database through a
binding name; there is one archive here, and 218 uses of `$data_binding=`
across ten templates all resolved to it.

Forty settings became `Option`s in `skins/deck/options.py`. A skin now
declares what an operator may set, the way a driver does, and the admin
page builds the form from it. `--explain` lists every driver, feed and
export, which it did not before.

`planets.py` is Meeus 32/33 on VSOP87, trimmed from 30,000 terms to 2,000
by `tools/vsop87trim.py`, which measures what the cut cost: six arcseconds
over two centuries. Pluto is Meeus 37. Against pyephem at six places over
forty years: three arcseconds of position, two seconds of rising, setting
and culmination. `$almanac.hasExtras` is now honestly true.

The ruff rule set is named in `pyproject.toml` rather than left to the
default, which grows between releases. What is not checked is listed with
its reason. `dateTime` is a column
name. An isolation boundary catches everything by design. The schema comes
from the file, so column names are interpolated into SQL. The Ecowitt
driver keeps its formatting because
`weewx-ecowitt` is still its own repository.

Fixed, each one measured rather than assumed:

* `stale_age` skipped a page whose template had changed. It means "the
  data may age", not "never rebuild this". A page is now stale when it is
  older than the skin it came from.
* A page whose template was deleted stayed on the site and kept being
  exported. Swept now, but never after a failed render.
* `%x`, `%X` and `calendar.month_abbr` follow the process locale, which
  the image does not set. A German page dated itself `08/27/26`.
* Six `Value`s never received the target, so the skin's formats and the
  language did not reach `$current.dateTime` or `$day.start`.
* `skinkit.ValueHelper` did not convert. A skin that adds up rain itself
  showed `1.11 in` on a page in millimetres.
* `Almanac.__getattr__` passed `self.almanac`, which does not exist, so
  every planet was handed an `Unknown` as its almanac.
* `RainTags` read `archive_day_rain` without checking for None. An archive
  whose daily summaries are not built cost twelve pages.
* `fake_get_report_years` returned integers where `$SummaryByYear` holds
  strings, and the templates concatenate.
* `min()` on an empty list, twice: a reading the station never had.
* A wind direction axis labelled every 45 degrees over sixteen ordinals.
* Skin colours were chosen against white. `#161616` is invisible on the
  dark theme, so a colour is now measured against what it is drawn on.
* Tables wider than the screen were cut off, not scrollable.
* A day-per-square calendar over ten years is 3650 squares and 6680 pixels
  on a phone. Over fourteen months it becomes months by years.

+++ src/weewx_evo/feeds/imagegenerator/canvas.py
@@ -101,7 +101,7 @@ class Canvas:
         self.draw.line(
             [(self._s(x), self._s(y)) for x, y in points],
             fill=theming.rgba(color, opacity),
-            width=max(1, int(round(self._s(width)))), joint="curve")
+            width=max(1, round(self._s(width))), joint="curve")
 
     def polygon(self, points: list[tuple[float, float]], color: str,
                 opacity: float = 1.0) -> None:
@@ -124,7 +124,7 @@ class Canvas:
         self.draw.ellipse([self._s(x) - r, self._s(y) - r,
                            self._s(x) + r, self._s(y) + r],
                           outline=theming.rgba(color),
-                          width=max(1, int(round(self._s(width)))))
+                          width=max(1, round(self._s(width))))
 
     def fade_under(self, points: list[tuple[float, float]], baseline: float,
                    color: str, opacity: float) -> None:
@@ -167,7 +167,7 @@ class Canvas:
         # costs one draw call per strip and looks no different.
         column = Image.new("L", (1, h))
         column.putdata([
-            int(round(255 * opacity * (1.0 - i / max(1, h - 1)) ** 1.4))
+            round(255 * opacity * (1.0 - i / max(1, h - 1)) ** 1.4)
             for i in range(h)])
         self.big.paste(theming.rgba(color)[:3], (x0, y0),
                        ImageChops.multiply(shape, column.resize((w, h))))
@@ -197,8 +197,8 @@ class Canvas:
         # at most and this is two allocations rather than one per column.
         row = Image.new("L", (w, 1))
         row.putdata([
-            int(round(255 * ((i / max(1, w - 1)) if towards_right
-                             else 1.0 - i / max(1, w - 1))))
+            round(255 * ((i / max(1, w - 1)) if towards_right
+                             else 1.0 - i / max(1, w - 1)))
             for i in range(w)])
         self.big.paste(theming.rgba(color)[:3], (left, top),
                        row.resize((w, h)))
@@ -231,7 +231,7 @@ class Canvas:
             try:
                 self._fonts[key] = (ImageFont.truetype(path, size)
                                     if path else _default_font(size))
-            except Exception:  # noqa: BLE001
+            except Exception:
                 self._fonts[key] = _default_font(size)
         return self._fonts[key]
 
@@ -316,7 +316,7 @@ def time_ticks(start: float, stop: float, wanted: int = 6,
 
     # Months and years. Stepped through the calendar, because adding
     # 2592000 seconds twelve times lands in the middle of December.
-    months = max(1, int(round(rough / (30 * 86400))))
+    months = max(1, round(rough / (30 * 86400)))
     if months >= 12:
         return _calendar_ticks(start, stop, years=max(1, months // 12),
                                language=language)
@@ -406,5 +406,5 @@ def label_for(value: float, step: float) -> str:
     """
     if step >= 1:
         return f"{value:.0f}"
-    places = max(0, min(4, int(math.ceil(-math.log10(step)))))
+    places = max(0, min(4, math.ceil(-math.log10(step))))
     return f"{value:.{places}f}"
